+++ test/service.js
@@ -3,10 +3,10 @@ const {Worker} = require('worker_threads');
 
 exec('redis-server');
 
-const auth = new Worker('./test/auth.js');
 const email = new Worker('./test/email.js');
 const logger = new Worker('./test/logger.js');
 const storeBuilder = new Worker('./test/storeBuilder.js');
+const auth = new Worker('./test/auth.js');
 
 setTimeout(() => {
     exec('./test/request.sh');