+++ CHANGELOG.md
@@ -4,3 +4,4 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 - Added feature for optional callbacks
+- Added RabbitMq support, option can be configured in the config file